materials-discovery/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "materials-discovery"
version = "0.1.0"
description = "No-DFT materials discovery scaffold for vZome"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "numpy>=1.26.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "types-PyYAML>=6.0.12.20240917",
]
mlip = [
  "ase>=3.23.0",
  "pymatgen>=2024.8.9",
  "spglib>=2.4.0",
  "phonopy>=2.32.0",
  "pycalphad>=0.11.0",
  "chgnet>=0.3.0",
  "mace-torch>=0.3.9",
]
analysis = [
  "pandas>=2.2.2",
  "pyarrow>=17.0.0",
]

[project.scripts]
mdisc = "materials_discovery.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/materials_discovery"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008", "B904"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_any_generics = true
show_error_codes = true
pretty = true
